Create Microservice

+++ Cut-Roll-News/src/Cut-Roll-News.Infrastructure/NewsLikes/Services/NewsLikeService.cs
@@ -1,15 +1,21 @@
 using Cut_Roll_News.Core.NewsArticles.Models;
+using Cut_Roll_News.Core.NewsArticles.Services;
 using Cut_Roll_News.Core.NewsLikes.Models;
 using Cut_Roll_News.Core.NewsLikes.Repositories;
 using Cut_Roll_News.Core.NewsLikes.Services;
+using Cut_Roll_News.Core.Users.Services;
 
 namespace Cut_Roll_News.Infrastructure.NewsLikes.Services;
 
 public class NewsLikeService : INewsLikeService
 {
     private readonly INewsLikeRepository _newsLikeRepository;
-    public NewsLikeService(INewsLikeRepository newsLikeRepository)
+    private readonly INewsArticleService _newsArticleService;
+    private readonly IUserService _userService;
+    public NewsLikeService(INewsLikeRepository newsLikeRepository, IUserService userService, INewsArticleService newsArticleService)
     {
+        _userService = userService ?? throw new ArgumentNullException(nameof(userService));
+        _newsArticleService = newsArticleService ?? throw new ArgumentNullException(nameof(newsArticleService));
         _newsLikeRepository = newsLikeRepository ?? throw new ArgumentNullException(nameof(newsLikeRepository));
     }
 
@@ -19,6 +25,14 @@ public class NewsLikeService : INewsLikeService
             throw new ArgumentNullException(nameof(userId));
         if (articleId == null)
             throw new ArgumentNullException(nameof(articleId));
+
+        var user = await _userService.GetUserByIdAsync(userId);
+        if (user == null)
+            throw new ArgumentException($"there is no user with id: {userId}");
+            
+        var news = await _newsArticleService.GetArticleAsNoTrackingAsync(articleId);
+        if (news == null)
+            throw new ArgumentException($"there is no user with id: {articleId}");
 
         return await _newsLikeRepository.CreateAsync(new NewsLike
         {
@@ -33,9 +47,21 @@ public class NewsLikeService : INewsLikeService
             throw new ArgumentNullException(nameof(userId));
         if (articleId == null)
             throw new ArgumentNullException(nameof(articleId));
+        
+        var user = await _userService.GetUserByIdAsync(userId);
+        if (user == null)
+            throw new ArgumentException($"there is no user with id: {userId}");
+            
+        var news = await _newsArticleService.GetArticleAsNoTrackingAsync(articleId);
+        if (news == null)
+            throw new ArgumentException($"there is no user with id: {articleId}");
+
+        var like = await _newsLikeRepository.GetByUserIdAndArticleId(userId, articleId.Value);
+        if (like == null)
+            throw new ArgumentException($"there is no like to delete");
 
         return await _newsLikeRepository.DeleteByUserIdAndArticleId(userId, articleId.Value)
-                ?? throw new InvalidOperationException($"Failed to delete like for user: {userId} and article: {articleId}");
+                    ?? throw new InvalidOperationException($"Failed to delete like for user: {userId} and article: {articleId}");
     }
     public async Task<NewsLike?> GetLikeByUserIdAndArticleId(string? userId, Guid? articleId)
     {
@@ -43,6 +69,14 @@ public class NewsLikeService : INewsLikeService
             throw new ArgumentNullException(nameof(userId));
         if (articleId == null)
             throw new ArgumentNullException(nameof(articleId));
+
+        var user = await _userService.GetUserByIdAsync(userId);
+        if (user == null)
+            throw new ArgumentException($"there is no user with id: {userId}");
+            
+        var news = await _newsArticleService.GetArticleAsNoTrackingAsync(articleId);
+        if (news == null)
+            throw new ArgumentException($"there is no user with id: {articleId}");
             
         return await _newsLikeRepository.GetByUserIdAndArticleId(userId, articleId.Value);
     }
@@ -52,6 +86,11 @@ public class NewsLikeService : INewsLikeService
         if (string.IsNullOrEmpty(userId))
             throw new ArgumentNullException(nameof(userId));
 
+        var user = await _userService.GetUserByIdAsync(userId);
+        if (user == null)
+            throw new ArgumentException($"there is no user with id: {userId}");
+            
+
         return await _newsLikeRepository.GetLikedNewsByUserIdAsync(userId);
     }
 
@@ -59,7 +98,11 @@ public class NewsLikeService : INewsLikeService
     {
         if (articleId == null)
             throw new ArgumentNullException(nameof(articleId));
-        
+            
+        var news = await _newsArticleService.GetArticleAsNoTrackingAsync(articleId);
+        if (news == null)
+            throw new ArgumentException($"there is no user with id: {articleId}");
+
         return await _newsLikeRepository.GetLikesCountByArticleIdAsync(articleId.Value);
     }
 
@@ -69,6 +112,14 @@ public class NewsLikeService : INewsLikeService
             throw new ArgumentNullException(nameof(userId));
         if (articleId == null)
             throw new ArgumentNullException(nameof(articleId));
+
+        var user = await _userService.GetUserByIdAsync(userId);
+        if (user == null)
+            throw new ArgumentException($"there is no user with id: {userId}");
+            
+        var news = await _newsArticleService.GetArticleAsNoTrackingAsync(articleId);
+        if (news == null)
+            throw new ArgumentException($"there is no user with id: {articleId}");
 
         return await _newsLikeRepository.ExistsAsync(userId, articleId.Value);
     }
